# erp-domain/openapi.yaml
openapi: 3.0.1
info:
  title: BigBright ERP Domain API
  description: Core APIs for multi-company ERP platform
  license:
    name: Proprietary
  version: v1
...
servers:
- url: http://localhost:51904
  description: Generated server url
paths:
...
  /api/v1/hr/employees/{id}:
    put:
      tags:
      - hr-controller
      operationId: updateEmployee
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseEmployeeDto"
    delete:
      tags:
      - hr-controller
      operationId: deleteEmployee
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /api/v1/payroll/runs:
    get:
      tags:
      - hr-payroll-controller
      operationId: listPayrollRuns
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListPayrollRunDto"
    post:
      tags:
      - hr-payroll-controller
      operationId: createPayrollRun
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayrollRunRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePayrollRunDto"
...
  /api/v1/hr/payroll-runs:
    get:
      tags:
      - hr-controller
      operationId: payrollRuns
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListPayrollRunDto"
    post:
      tags:
      - hr-controller
      operationId: createPayrollRun_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollRunRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePayrollRunDto"
  /api/v1/hr/leave-requests:
    get:
      tags:
      - hr-controller
      operationId: leaveRequests
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListLeaveRequestDto"
      deprecated: true
    post:
      tags:
      - hr-controller
      operationId: createLeaveRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLeaveRequestDto"
      deprecated: true
  /api/v1/hr/employees:
    get:
      tags:
      - hr-controller
      operationId: employees
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListEmployeeDto"
    post:
      tags:
      - hr-controller
      operationId: createEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseEmployeeDto"
  /api/v1/hr/attendance/mark/{employeeId}:
    post:
      tags:
      - hr-controller
      operationId: markAttendance
      parameters:
      - name: employeeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAttendanceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAttendanceDto"
  /api/v1/hr/attendance/bulk-mark:
    post:
      tags:
      - hr-controller
      operationId: bulkMarkAttendance
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkMarkAttendanceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListAttendanceDto"
...
  /api/v1/hr/leave-requests/{id}/status:
    patch:
      tags:
      - hr-controller
      operationId: updateLeaveStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseLeaveRequestDto"
      deprecated: true
...
  /api/v1/hr/attendance/employee/{employeeId}:
    get:
      tags:
      - hr-controller
      operationId: employeeAttendance
      parameters:
      - name: employeeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: startDate
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: endDate
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListAttendanceDto"
  /api/v1/hr/attendance/date/{date}:
    get:
      tags:
      - hr-controller
      operationId: attendanceByDate
      parameters:
      - name: date
        in: path
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListAttendanceDto"
...
components:
  schemas:
...
    EmployeeRequest:
      required:
      - email
      - firstName
      - lastName
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        hiredDate:
          type: string
          format: date
        employeeType:
          type: string
        paymentSchedule:
          type: string
        monthlySalary:
          type: number
        dailyWage:
          type: number
        workingDaysPerMonth:
          type: integer
          format: int32
        weeklyOffDays:
          type: integer
          format: int32
        standardHoursPerDay:
          type: number
        overtimeRateMultiplier:
          type: number
        doubleOtRateMultiplier:
          type: number
        bankAccountNumber:
          type: string
        bankName:
          type: string
        ifscCode:
          type: string
    ApiResponseEmployeeDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/EmployeeDto"
        timestamp:
          type: string
          format: date-time
    EmployeeDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        hiredDate:
          type: string
          format: date
...
    CreatePayrollRunRequest:
      type: object
      properties:
        runType:
          type: string
          enum:
          - WEEKLY
          - MONTHLY
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        remarks:
          type: string
    ApiResponsePayrollRunDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/PayrollRunDto"
        timestamp:
          type: string
          format: date-time
    PayrollRunDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        runNumber:
          type: string
        runType:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        status:
          type: string
        totalEmployees:
          type: integer
          format: int32
        totalPresentDays:
          type: number
        totalOvertimeHours:
          type: number
        totalBasePay:
          type: number
        totalOvertimePay:
          type: number
        totalDeductions:
          type: number
        totalNetPay:
          type: number
        journalEntryId:
          type: integer
          format: int64
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        approvedBy:
          type: string
        approvedAt:
          type: string
          format: date-time
        postedBy:
          type: string
        postedAt:
          type: string
          format: date-time
        remarks:
          type: string
    PayrollRunRequest:
      required:
      - creditAccountId
      - debitAccountId
      - initiatedBy
      - payrollDate
      - postingAmount
      type: object
      properties:
        payrollDate:
          type: string
          format: date
        initiatedBy:
          type: string
        debitAccountId:
          type: integer
          format: int64
        creditAccountId:
          type: integer
          format: int64
        postingAmount:
          type: number
...
    LeaveRequestRequest:
      required:
      - endDate
      - leaveType
      - startDate
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        leaveType:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
        status:
          type: string
    ApiResponseLeaveRequestDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/LeaveRequestDto"
        timestamp:
          type: string
          format: date-time
    LeaveRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        leaveType:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
    MarkAttendanceRequest:
      required:
      - status
      type: object
      properties:
        date:
          type: string
          format: date
        status:
          type: string
        checkInTime:
          $ref: "#/components/schemas/LocalTime"
        checkOutTime:
          $ref: "#/components/schemas/LocalTime"
        regularHours:
          type: number
        overtimeHours:
          type: number
        doubleOvertimeHours:
          type: number
        holiday:
          type: boolean
        weekend:
          type: boolean
        remarks:
          type: string
    ApiResponseAttendanceDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/AttendanceDto"
        timestamp:
          type: string
          format: date-time
    AttendanceDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        employeeType:
          type: string
        date:
          type: string
          format: date
        status:
          type: string
        checkInTime:
          $ref: "#/components/schemas/LocalTime"
        checkOutTime:
          $ref: "#/components/schemas/LocalTime"
        regularHours:
          type: number
        overtimeHours:
          type: number
        doubleOvertimeHours:
          type: number
        holiday:
          type: boolean
        weekend:
          type: boolean
        remarks:
          type: string
        markedBy:
          type: string
        markedAt:
          type: string
          format: date-time
    BulkMarkAttendanceRequest:
      required:
      - date
      - employeeIds
      - status
      type: object
      properties:
        employeeIds:
          type: array
          items:
            type: integer
            format: int64
        date:
          type: string
          format: date
        status:
          type: string
        checkInTime:
          $ref: "#/components/schemas/LocalTime"
        checkOutTime:
          $ref: "#/components/schemas/LocalTime"
        regularHours:
          type: number
        overtimeHours:
          type: number
        remarks:
          type: string
    ApiResponseListAttendanceDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceDto"
        timestamp:
          type: string
          format: date-time
...
    StatusRequest:
      type: object
      properties:
        status:
          type: string
...
    ApiResponseListPayrollRunDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/PayrollRunDto"
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListLeaveRequestDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/LeaveRequestDto"
        timestamp:
          type: string
          format: date-time
    ApiResponseListEmployeeDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeDto"
        timestamp:
          type: string
          format: date-time
...